// src/app/types.d.ts
export type Movie = {
  Title: string
  Year: string
  imdbID: string
  Type: MovieType
  Poster: string
}

export enum MovieType {
  Movie = 'movie',
  Series = 'series'
}

export type MovieDetails = {
  Title: string
  Year: string
  Rated: string
  Released: string
  Runtime: string
  Genre: string
  Director: string
  Writer: string
  Actors: string
  Plot: string
  Language: string
  Country: string
  Awards: string
  Poster: string
  Ratings: Rating[]
  Metascore: string
  imdbRating: string
  imdbVotes: string
  imdbID: string
  Type: string
  DVD: string
  BoxOffice: string
  Production: string
  Website: string
  Response: string
}

export type Rating = {
  Source: string
  Value: string
}